Added plantUML diagrams to README
Currently I feel only a sequence and a use case diagram is required to understand the program - hence that is all I have included. The tokens at the end of the links have to be updated when the files update due to this being a private repository as of now.